refactor: introduce RfcMetadata model
Creates a new model to store metadata about RFCs, replacing the previous AreaAssignment model and better capturing the source of an RFC as represented in the system this is replacing.

It is expected that as new RFCs come into existance, the RPC publishing process will tell this system about the RFC at the same time it tells the datatracker.